Changes in the dashboard controller. Export dashboard access code is moved to dashboard export controller
