+++ apps/acqdat_api/lib/acqdat_api_web/controllers/dashboard_management/dashboard_controller.ex
@@ -191,58 +191,7 @@ defmodule AcqdatApiWeb.DashboardManagement.DashboardController do
     end
   end
 
-  def exported_dashboard(conn, params) do
-    case conn.status do
-      nil ->
-        exported_dashboard = conn.assigns.exported_dashboard
-
-        dashboard =
-          check_exported_dashboard(exported_dashboard.is_secure, params, exported_dashboard)
-
-        case dashboard do
-          {:ok, dashboard} ->
-            conn
-            |> put_status(200)
-            |> render("show.json", %{dashboard: dashboard})
-
-          {:error, message} ->
-            send_error(conn, 400, message)
-
-          nil ->
-            conn
-            |> send_error(401, "Unauthorized link")
-        end
-
-      401 ->
-        conn
-        |> send_error(401, "Unauthorized link")
-    end
-  end
-
   ############################# private functions ###########################
-
-  defp check_exported_dashboard(true, params, exported_dashboard) do
-    case check_password(params["password"], exported_dashboard.password) do
-      false ->
-        nil
-
-      true ->
-        Dashboard.get_by_uuid(exported_dashboard.dashboard_uuid)
-    end
-  end
-
-  defp check_exported_dashboard(false, params, exported_dashboard) do
-    Dashboard.get_by_uuid(exported_dashboard.dashboard_uuid)
-  end
-
-  defp check_password(password, db_password) do
-    if password == db_password do
-      true
-    else
-      false
-    end
-  end
-
   defp add_avatar_to_params(conn, params) do
     params = Map.put(params, "avatar", "")
