+++ apps/acqdat_api/lib/acqdat_api_web/router.ex
@@ -33,7 +33,7 @@ defmodule AcqdatApiWeb.Router do
 
     get(
       "/dashboards/:dashboard_uuid/verify",
-      DashboardManagement.DashboardController,
+      DashboardManagement.DashboardExportController,
       :exported_dashboard
     )
