+++ mix.exs
@@ -73,7 +73,7 @@ defmodule AcqdatUmbrella.MixProject do
       {:gen_retry, "~> 1.2.0"},
 
       # testing
-      {:excoveralls, "~> 0.10", only: :test},
+      {:excoveralls, "~> 0.10", only: :test}
     ]
   end
